+++ tests/unit/Composer/Read/Register/BitReadRegisterAddressTest.php
@@ -1,29 +1,29 @@
 <?php
 
-namespace Tests\unit\Composer\Read;
+namespace Tests\unit\Composer\Read\Register;
 
 
 use ModbusTcpClient\Packet\ModbusFunction\ReadHoldingRegistersResponse;
-use ModbusTcpClient\Composer\Read\BitReadAddress;
+use ModbusTcpClient\Composer\Read\Register\BitReadRegisterAddress;
 use PHPUnit\Framework\TestCase;
 
-class BitAddressTest extends TestCase
+class BitReadRegisterAddressTest extends TestCase
 {
     public function testGetSize()
     {
-        $address = new BitReadAddress(1, 0);
+        $address = new BitReadRegisterAddress(1, 0);
         $this->assertEquals(1, $address->getSize());
     }
 
     public function testGetName()
     {
-        $address = new BitReadAddress(1, 0, 'alarm1_do');
+        $address = new BitReadRegisterAddress(1, 0, 'alarm1_do');
         $this->assertEquals('alarm1_do', $address->getName());
     }
 
     public function testDefaultGetName()
     {
-        $address = new BitReadAddress(1, 1);
+        $address = new BitReadRegisterAddress(1, 1);
         $this->assertEquals('bit_1_1', $address->getName());
     }
 
@@ -31,16 +31,16 @@ class BitAddressTest extends TestCase
     {
         $responsePacket = new ReadHoldingRegistersResponse("\x02\x00\x05", 3, 33152);
 
-        $this->assertTrue((new BitReadAddress(0, 0))->extract($responsePacket));
-        $this->assertFalse((new BitReadAddress(0, 1))->extract($responsePacket));
-        $this->assertTrue((new BitReadAddress(0, 2))->extract($responsePacket));
+        $this->assertTrue((new BitReadRegisterAddress(0, 0))->extract($responsePacket));
+        $this->assertFalse((new BitReadRegisterAddress(0, 1))->extract($responsePacket));
+        $this->assertTrue((new BitReadRegisterAddress(0, 2))->extract($responsePacket));
     }
 
     public function testExtractWithCallback()
     {
         $responsePacket = new ReadHoldingRegistersResponse("\x02\x00\x05", 3, 33152);
 
-        $address = new BitReadAddress(0, 0, 'name', function ($value) {
+        $address = new BitReadRegisterAddress(0, 0, 'name', function ($value) {
             return 'prefix_' . $value; // transform value after extraction
         });
         $this->assertEquals('prefix_1', $address->extract($responsePacket));