+++ tests/unit/Composer/Read/Register/ReadRegisterAddressTest.php
@@ -1,21 +1,21 @@
 <?php
 
-namespace Tests\unit\Composer\Read;
+namespace Tests\unit\Composer\Read\Register;
 
 
 use ModbusTcpClient\Composer\Address;
-use ModbusTcpClient\Composer\Read\ReadAddress;
+use ModbusTcpClient\Composer\Read\Register\ReadRegisterAddress;
 use ModbusTcpClient\Packet\ModbusFunction\ReadHoldingRegistersResponse;
 use PHPUnit\Framework\TestCase;
 
-class ReadAddressTest extends TestCase
+class ReadRegisterAddressTest extends TestCase
 {
     /**
      * @dataProvider sizeProvider
      */
     public function testGetSize($type, $expectedSize)
     {
-        $address = new ReadAddress(1, $type);
+        $address = new ReadRegisterAddress(1, $type);
         $this->assertEquals($expectedSize, $address->getSize());
     }
 
@@ -36,24 +36,24 @@ class ReadAddressTest extends TestCase
      */
     public function testInvalidType()
     {
-        new ReadAddress(1, Address::TYPE_BYTE);
+        new ReadRegisterAddress(1, Address::TYPE_BYTE);
     }
 
     public function testGetName()
     {
-        $address = new ReadAddress(1, Address::TYPE_INT16, 'temp1');
+        $address = new ReadRegisterAddress(1, Address::TYPE_INT16, 'temp1');
         $this->assertEquals('temp1', $address->getName());
     }
 
     public function testDefaultGetName()
     {
-        $address = new ReadAddress(1, Address::TYPE_INT16);
+        $address = new ReadRegisterAddress(1, Address::TYPE_INT16);
         $this->assertEquals('int16_1', $address->getName());
     }
 
     public function testGetAddress()
     {
-        $address = new ReadAddress(1, Address::TYPE_INT16);
+        $address = new ReadRegisterAddress(1, Address::TYPE_INT16);
         $this->assertEquals(1, $address->getAddress());
     }
 
@@ -63,7 +63,7 @@ class ReadAddressTest extends TestCase
 
         $cbAddress = null;
         $cbResponse = null;
-        $address = new ReadAddress(
+        $address = new ReadRegisterAddress(
             1,
             Address::TYPE_UINT16,
             null,
@@ -87,7 +87,7 @@ class ReadAddressTest extends TestCase
 
         $errCbAddress = null;
         $errCbResponse = null;
-        $address = new ReadAddress(
+        $address = new ReadRegisterAddress(
             0,
             Address::TYPE_UINT64,
             null,
@@ -119,7 +119,7 @@ class ReadAddressTest extends TestCase
 
         $errCbAddress = null;
         $errCbResponse = null;
-        $address = new ReadAddress(
+        $address = new ReadRegisterAddress(
             0,
             Address::TYPE_INT16,
             null,
@@ -140,10 +140,31 @@ class ReadAddressTest extends TestCase
         $this->assertEquals($address, $errCbAddress);
     }
 
+
+    /**
+     * @expectedException \RuntimeException
+     * @expectedExceptionMessage test
+     */
+    public function testExtractWithNoErrorCallback()
+    {
+        $responsePacket = new ReadHoldingRegistersResponse("\x08\x00\x00\x00\x00\x00\x00\x00\x00", 3, 33152);
+
+        $address = new ReadRegisterAddress(
+            0,
+            Address::TYPE_INT16,
+            null,
+            function () {
+                throw new \RuntimeException('test');
+            }
+        );
+
+        $address->extract($responsePacket);
+    }
+
     public function testExtractInt16()
     {
         $responsePacket = new ReadHoldingRegistersResponse("\x08\x00\x01\x80\x00\x00\x03\x00\x04", 3, 33152);
-        $address = new ReadAddress(1, Address::TYPE_INT16);
+        $address = new ReadRegisterAddress(1, Address::TYPE_INT16);
 
         $value = $address->extract($responsePacket);
 
@@ -153,7 +174,7 @@ class ReadAddressTest extends TestCase
     public function testExtractUint16()
     {
         $responsePacket = new ReadHoldingRegistersResponse("\x08\x00\x01\x80\x00\x00\x03\x00\x04", 3, 33152);
-        $address = new ReadAddress(1, Address::TYPE_UINT16);
+        $address = new ReadRegisterAddress(1, Address::TYPE_UINT16);
 
         $value = $address->extract($responsePacket);
 
@@ -163,7 +184,7 @@ class ReadAddressTest extends TestCase
     public function testExtractUint32()
     {
         $responsePacket = new ReadHoldingRegistersResponse("\x08\x00\x01\xFF\xFF\x7F\xFF\x00\x04", 3, 33152);
-        $address = new ReadAddress(1, Address::TYPE_UINT32);
+        $address = new ReadRegisterAddress(1, Address::TYPE_UINT32);
 
         $value = $address->extract($responsePacket);
 
@@ -173,7 +194,7 @@ class ReadAddressTest extends TestCase
     public function testExtractInt32()
     {
         $responsePacket = new ReadHoldingRegistersResponse("\x08\x00\x01\x00\x00\x80\x00\x00\x04", 3, 33152);
-        $address = new ReadAddress(1, Address::TYPE_INT32);
+        $address = new ReadRegisterAddress(1, Address::TYPE_INT32);
 
         $value = $address->extract($responsePacket);
 
@@ -183,7 +204,7 @@ class ReadAddressTest extends TestCase
     public function testExtractFloat()
     {
         $responsePacket = new ReadHoldingRegistersResponse("\x08\x00\x01\xAA\xAB\x3F\x2A\x00\x04", 3, 33152);
-        $address = new ReadAddress(1, Address::TYPE_FLOAT);
+        $address = new ReadRegisterAddress(1, Address::TYPE_FLOAT);
 
         $value = $address->extract($responsePacket);
 
@@ -197,7 +218,7 @@ class ReadAddressTest extends TestCase
         }
 
         $responsePacket = new ReadHoldingRegistersResponse("\x08\xFF\xFF\x7F\xFF\x00\x00\x00\x00", 3, 33152);
-        $address = new ReadAddress(0, Address::TYPE_UINT64);
+        $address = new ReadRegisterAddress(0, Address::TYPE_UINT64);
 
         $value = $address->extract($responsePacket);
 
@@ -211,7 +232,7 @@ class ReadAddressTest extends TestCase
         }
 
         $responsePacket = new ReadHoldingRegistersResponse("\x08\xFF\xFF\xFF\xFF\xFF\xFF\xFF\xFF", 3, 33152);
-        $address = new ReadAddress(0, Address::TYPE_INT64);
+        $address = new ReadRegisterAddress(0, Address::TYPE_INT64);
 
         $value = $address->extract($responsePacket);